GLab_validation/requirements.txt
python-gitlab
python-dotenv
requests

